+++ packages/ice/src/service/preBundleCJSDeps.ts
@@ -13,7 +13,7 @@ import flattenId from '../utils/flattenId.js';
 import formatPath from '../utils/formatPath.js';
 import { BUILDIN_CJS_DEPS, BUILDIN_ESM_DEPS } from '../constant.js';
 import type { DepScanData } from '../esbuild/scan.js';
-import aliasPlugin from '../esbuild/alias.js';
+import resolvePlugin from '../esbuild/resolve.js';
 import emptyCSSPlugin from '../esbuild/emptyCSS.js';
 import cssModulesPlugin from '../esbuild/cssModules.js';
 import escapeLocalIdent from '../utils/escapeLocalIdent.js';
@@ -96,7 +96,7 @@ export default async function preBundleCJSDeps(options: PreBundleDepsOptions): P
       ignoreAnnotations: true,
       plugins: [
         emptyCSSPlugin(),
-        aliasPlugin({ alias, format: 'cjs', externalDependencies: false }),
+        resolvePlugin({ alias, format: 'cjs', externalDependencies: false }),
         cssModulesPlugin({
           extract: false,
           generateLocalIdentName: function (name: string, filename: string) {