+++ packages/webpack-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
-## v3.0.0
+## 1.0.1
 
-- [feat] basic webpack configuration
+- [fix] Lock version of `@ice/bundles`.
+- [feat] Support redirect imports.
+- [refactor] Move @swc/core, core-js and all swc plugins into `@ice/bundles`, to lock the version of them.
+
+## 1.0.0
+
+- [feat] Basic webpack configuration.